src/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain_community
faiss-cpu
pypdfium2
pytesseract
pillow
requests
mysql-connector-python
SQLAlchemy
pandas
python-dotenv
streamlit

